requirements.txt
fastapi
uvicorn[standard]
alpaca-py
pandas
fastapi
uvicorn[standard]
alpaca-py
pandas
python-dotenv
aiohttp
PyYAML
requests-mock
pytest-asyncio
